+++ src/nexgen/__main__.py
@@ -5,6 +5,6 @@ def main(args=None):
     version_parser.parse_args(args)
 
 
-# Test with python -m nexgen
+# Test with python -m nexgen -V
 if __name__ == "__main__":
     main()